Create UI to issue citations
Created fxml file 'issue-citation.fxml'. This page assumes a driver has already been selected from the LoggedIn page. Users are able to select a citation via combobox, can enter a citation amount, and can click the issue button to create the citation for a user.

Currently, the UI does not resize when the window size is changed. This will be optomized in a later commit.

Main currently updated to open the issue-citation.fxml file when ran.