Update UI for log-in, logged-in and issue-citation pages.
Modified issue-citation.fxml to dynamically resize contents when window size is changed. Placed body of page in the center of the border pane. Deleted unnecessary hboxes and vboxes. Changed alignment settings.

Changed log-in.fxml's password entry from a text field to a password field to mask characters when a password is typed.

Changed text field colors to white on logged-in.fxml for increased readability.